requirements.txt
pyserial
RPi.GPIO
pynmea2
